[BEAM-12513] Reformat code to be easier to find.
