--- sdks/go/examples/snippets/combine_test.go
@@ -1,37 +1,0 @@
-// Licensed to the Apache Software Foundation (ASF) under one or more
-// contributor license agreements.  See the NOTICE file distributed with
-// this work for additional information regarding copyright ownership.
-// The ASF licenses this file to You under the Apache License, Version 2.0
-// (the "License"); you may not use this file except in compliance with
-// the License.  You may obtain a copy of the License at
-//
-//    http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package snippets
-
-import (
-	"testing"
-
-	"github.com/apache/beam/sdks/go/pkg/beam/testing/passert"
-	"github.com/apache/beam/sdks/go/pkg/beam/testing/ptest"
-)
-
-func TestCombineWithDefault_useDefault(t *testing.T) {
-	p, s, input := ptest.CreateList([]int{})
-	avg := globallyAverageWithDefault(s, input)
-	passert.Equals(s, avg, float64(0))
-	ptest.RunAndValidate(t, p)
-}
-
-func TestCombineWithDefault_useAverage(t *testing.T) {
-	p, s, input := ptest.CreateList([]int{1, 2, 3})
-	avg := globallyAverageWithDefault(s, input)
-	passert.Equals(s, avg, float64(2.0))
-	ptest.RunAndValidate(t, p)
-}
